Move features_from_csv into features module
