+++ packages/vm/src/lib.rs
@@ -16,7 +16,7 @@ mod serde;
 pub mod testing;
 mod traits;
 
-pub use crate::cache::{features_from_csv, CosmCache};
+pub use crate::cache::CosmCache;
 pub use crate::calls::{
     call_handle, call_handle_raw, call_init, call_init_raw, call_query, call_query_raw,
 };
@@ -25,6 +25,7 @@ pub use crate::errors::{
     make_ffi_bad_argument, make_ffi_foreign_panic, make_ffi_other, make_ffi_out_of_gas, FfiError,
     FfiResult, VmError, VmResult,
 };
+pub use crate::features::features_from_csv;
 pub use crate::instance::Instance;
 pub use crate::modules::FileSystemCache;
 pub use crate::serde::{from_slice, to_vec};
